Reduce code for extended metadata
Some of the code used for rendering extended
metadata is very similar to the code used for
rendering common metadata. This patch
attempts to reduce the resulting code duplication
a bit.

Includes a logic change on how we store metadata
values in formik, by storing them the same
way extended metadata is stored.
