+++ src/components/events/partials/ModalTabsAndPages/DetailsMetadataTab.tsx
@@ -5,52 +5,54 @@ import { Field } from "../../../shared/Field";
 import cn from "classnames";
 import _ from "lodash";
 import Notifications from "../../../shared/Notifications";
-import RenderDate from "../../../shared/RenderDate";
 import RenderMultiField from "../../../shared/wizard/RenderMultiField";
 import RenderField from "../../../shared/wizard/RenderField";
 import { getUserInformation } from "../../../../selectors/userInfoSelectors";
-import { hasAccess } from "../../../../utils/utils";
+import {
+	hasAccess,
+} from "../../../../utils/utils";
 import { getMetadataCollectionFieldName } from "../../../../utils/resourceUtils";
 import { useAppDispatch, useAppSelector } from "../../../../store";
 import { MetadataCatalog } from "../../../../slices/eventSlice";
 import { AsyncThunk } from "@reduxjs/toolkit";
+import RenderDate from "../../../shared/RenderDate";
 
 /**
  * This component renders metadata details of a certain event or series
  */
 const DetailsMetadataTab = ({
-	metadataFields,
-	updateResource,
 	resourceId,
-	header,
+	metadata,
+	updateResource,
 	editAccessRole,
+	header,
 }: {
-	metadataFields: MetadataCatalog,
-	updateResource: AsyncThunk<void, { id: string; values: { [key: string]: any; }; }, any>
 	resourceId: string,
-	header: string,
+	metadata: MetadataCatalog[],
+	updateResource: AsyncThunk<void, {
+		id: string;
+		values: { [key: string]: any; };
+		catalog: MetadataCatalog;
+	}, any> //(id: string, values: { [key: string]: any }, catalog: MetadataCatalog) => void,
 	editAccessRole: string,
+	header?: string
 }) => {
 	const { t } = useTranslation();
 	const dispatch = useAppDispatch();
 
 	const user = useAppSelector(state => getUserInformation(state));
 
-	const handleSubmit = (values: { [key: string]: any }) => {
-		dispatch(updateResource({id: resourceId, values}));
+	const handleSubmit = (values: { [key: string]: any }, catalog: MetadataCatalog) => {
+		dispatch(updateResource({id: resourceId, values, catalog}));
 	};
 
 	// set current values of metadata fields as initial values
-	const getInitialValues = () => {
-		let initialValues: { [key: string]: string | string[] } = {};
+	const getInitialValues = (metadataCatalog: MetadataCatalog) => {
+		let initialValues: { [key: string]: any } = {};
 
 		// Transform metadata fields and their values provided by backend (saved in redux)
-		if (
-			!!metadataFields &&
-			!!metadataFields.fields &&
-			metadataFields.fields.length > 0
-		) {
-			metadataFields.fields.forEach((field) => {
+		if (!!metadataCatalog.fields && metadataCatalog.fields.length > 0) {
+			metadataCatalog.fields.forEach((field) => {
 				initialValues[field.id] = field.value;
 			});
 		}
@@ -58,7 +60,7 @@ const DetailsMetadataTab = ({
 		return initialValues;
 	};
 
-	const checkValidity = (formik: FormikProps<{ [key: string]: string | string[] }>) => {
+	const checkValidity = (formik: FormikProps<{}>) => {
 		if (formik.dirty && formik.isValid && hasAccess(editAccessRole, user)) {
 			// check if user provided values differ from initial ones
 			return !_.isEqual(formik.values, formik.initialValues);
@@ -68,106 +70,116 @@ const DetailsMetadataTab = ({
 	};
 
 	return (
-		// initialize form
-		<Formik
-			enableReinitialize
-			initialValues={getInitialValues()}
-			onSubmit={(values) => handleSubmit(values)}
-		>
-			{(formik) => (
-				<>
-					<div className="modal-content">
-						<div className="modal-body">
-							<Notifications context="not-corner" />
-							<div className="full-col">
-								<div className="obj tbl-list">
-									<header className="no-expand">{t(header)}</header>
-									<div className="obj-container">
-										<table className="main-tbl">
-											<tbody>
-												{/* Render table row for each metadata field depending on type */}
-												{!!metadataFields &&
-													!!metadataFields.fields &&
-													metadataFields.fields.map((field, key) => (
-														<tr key={key}>
-															<td>
-																<span>{t(field.label)}</span>
-																{field.required && (
-																	<i className="required">*</i>
-																)}
-															</td>
-															{field.readOnly ? (
-																// non-editable field if readOnly is set
-																!!field.collection &&
-																field.collection.length !== 0 ? (
-																	<td>{getMetadataCollectionFieldName(field, field, t)}</td>
-																) : (
-																	<td>{
-																		field.type === "time" || field.type === "date"
-																			? <RenderDate date={field.value as string} />
-																			: field.value
-																	}</td>
-																)
-															) : (
-																<td className="editable">
-																	{/* Render single value or multi value editable input */}
-																	{field.type === "mixed_text" &&
-																	field.collection?.length !== 0 ? (
-																		<Field
-																			name={field.id}
-																			fieldInfo={field}
-																			showCheck
-																			component={RenderMultiField}
-																		/>
+		<div className="modal-content">
+			<div className="modal-body">
+				{/* Notifications */}
+				<Notifications context="not_corner" />
+
+				<div className="full-col">
+					{
+						//iterate through metadata catalogs
+						!!metadata &&
+							metadata.length > 0 &&
+							metadata.map((catalog, key) => (
+								// initialize form
+								<Formik
+									enableReinitialize
+									initialValues={getInitialValues(catalog)}
+									onSubmit={(values) => handleSubmit(values, catalog)}
+								>
+									{(formik) => (
+										/* Render table for each metadata catalog */
+										<div className="obj tbl-details" key={key}>
+											<header>
+												<span>{t(header ? header : catalog.title)}</span>
+											</header>
+											<div className="obj-container">
+												<table className="main-tbl">
+													<tbody>
+														{/* Render table row for each metadata field depending on type */}
+														{!!catalog.fields &&
+															catalog.fields.map((field, index) => (
+																<tr key={index}>
+																	<td>
+																		<span>{t(field.label)}</span>
+																		{field.required && (
+																			<i className="required">*</i>
+																		)}
+																	</td>
+																	{field.readOnly ||
+																	!hasAccess(editAccessRole, user) ? (
+																		// non-editable field if readOnly is set or user doesn't have edit access rights
+																		!!field.collection &&
+																		field.collection.length !== 0 ? (
+																			<td>{getMetadataCollectionFieldName(field, field, t)}</td>
+																		) : (
+																			<td>{
+																				field.type === "time" || field.type === "date"
+																					? <RenderDate date={field.value as string} />
+																					: field.value
+																			}</td>
+																		)
 																	) : (
-																		<Field
-																			name={field.id}
-																			metadataField={field}
-																			showCheck
-																			component={RenderField}
-																		/>
+																		<td className="editable">
+																			{/* Render single value or multi value editable input */}
+																			{field.type === "mixed_text" &&
+																			field.collection?.length !== 0 ? (
+																				<Field
+																					name={field.id}
+																					fieldInfo={field}
+																					showCheck
+																					component={RenderMultiField}
+																				/>
+																			) : (
+																				<Field
+																					name={field.id}
+																					metadataField={field}
+																					showCheck
+																					component={RenderField}
+																				/>
+																			)}
+																		</td>
 																	)}
-																</td>
-															)}
-														</tr>
-													))}
-											</tbody>
-										</table>
-									</div>
+																</tr>
+															))}
+													</tbody>
+												</table>
+											</div>
 
-									{formik.dirty && (
-										<>
-											{/* Render buttons for updating metadata */}
-											<footer>
-												<button
-													type="submit"
-													onClick={() => formik.handleSubmit()}
-													disabled={!checkValidity(formik)}
-													className={cn("submit", {
-														active: checkValidity(formik),
-														inactive: !checkValidity(formik),
-													})}
-												>
-													{t("SAVE")}
-												</button>
-												<button
-													className="cancel"
-													onClick={() => formik.resetForm()}
-												>
-													{t("CANCEL")}
-												</button>
-											</footer>
+											{formik.dirty && (
+												<>
+													{/* Render buttons for updating metadata */}
+													<footer>
+														<button
+															type="submit"
+															onClick={() => formik.handleSubmit()}
+															disabled={!checkValidity(formik)}
+															className={cn("submit", {
+																active: checkValidity(formik),
+																inactive: !checkValidity(formik),
+															})}
+														>
+															{t("SAVE")}
+														</button>
+														<button
+															className="cancel"
+															onClick={() => formik.resetForm()}
+														>
+															{t("CANCEL")}
+														</button>
+													</footer>
 
-											<div className="btm-spacer" />
-										</>
+													<div className="btm-spacer" />
+												</>
+											)}
+										</div>
 									)}
-								</div>
-							</div>
-						</div>
-					</div>
-				</>
-			)}
-		</Formik>
+								</Formik>
+							))
+					}
+				</div>
+			</div>
+		</div>
 	);
 };
