+++ src/components/events/partials/modals/SeriesDetails.tsx
@@ -16,8 +16,7 @@ import SeriesDetailsAccessTab from "../ModalTabsAndPages/SeriesDetailsAccessTab"
 import SeriesDetailsThemeTab from "../ModalTabsAndPages/SeriesDetailsThemeTab";
 import SeriesDetailsStatisticTab from "../ModalTabsAndPages/SeriesDetailsStatisticTab";
 import SeriesDetailsFeedsTab from "../ModalTabsAndPages/SeriesDetailsFeedsTab";
-import DetailsMetadataTab from "../ModalTabsAndPages/DetailsMetadataTab";
-import DetailsExtendedMetadataTab from "../ModalTabsAndPages/DetailsExtendedMetadataTab";
+import DetailsExtendedMetadataTab from "../ModalTabsAndPages/DetailsMetadataTab";
 import { useAppDispatch, useAppSelector } from "../../../../store";
 import {
 	fetchSeriesStatistics,
@@ -127,12 +126,12 @@ const SeriesDetails = ({
 			{/* render modal content depending on current page */}
 			<div>
 				{page === 0 && (
-					<DetailsMetadataTab
-						metadataFields={metadataFields}
+					<DetailsExtendedMetadataTab
 						resourceId={seriesId}
-						header={tabs[page].tabNameTranslation}
+						metadata={[metadataFields]}
 						updateResource={updateSeriesMetadata}
 						editAccessRole="ROLE_UI_SERIES_DETAILS_METADATA_EDIT"
+						header={tabs[page].tabNameTranslation}
 					/>
 				)}
 				{page === 1 && (
