+++ src/slices/seriesSlice.ts
@@ -5,7 +5,6 @@ import {
 	getURLParams,
 	prepareAccessPolicyRulesForPost,
 	prepareSeriesExtendedMetadataFieldsForPost,
-	prepareSeriesMetadataFieldsForPost,
 	transformMetadataCollection,
 } from "../utils/resourceUtils";
 import {
@@ -192,10 +191,10 @@ export const postNewSeries = createAppAsyncThunk('series/postNewSeries', async (
 	const { values, metadataInfo, extendedMetadata } = params
 
 	// prepare metadata provided by user
-	let metadataFields = prepareSeriesMetadataFieldsForPost(
-		metadataInfo.fields,
+	let metadataFields = prepareSeriesExtendedMetadataFieldsForPost(
+		[metadataInfo],
 		values
-	);
+	)[0].fields;
 	let extendedMetadataFields = prepareSeriesExtendedMetadataFieldsForPost(
 		extendedMetadata,
 		values
